+++ pubspec.yaml
@@ -1,7 +1,8 @@
 name: font_scaler
 description: "A new Flutter package project."
-version: 0.0.1
-homepage:
+version: 1.0.0
+homepage: "https://github.com/Mirzaazmath"
+repository: "https://github.com/Mirzaazmath/font_scaler"
 
 environment:
   sdk: ^3.7.0
